requirements.txt
aiohttp==3.12.15
beautifulsoup4==4.14.2
openai==2.8.0
pandas==2.3.3
polars==1.34.0
python-dotenv==1.2.1
python-telegram-bot==22.5
requests==2.32.5
